requirements.txt
aiogram==3.*
python-docx
PyPDF2
